// src/app/core/models/thanh-cai.ts
export class ThanhCaiList {
  MAPMIS: string = '';
  TENTHANHCAI: string = '';
  TENCONGTY: string = '';
  TRUYENTAIDIEN: string = '';
  CAPDA: string = '';
  TENTRAM: string = '';
}

export class ThanhCaiDetail {
  Mapmis: string | null | undefined;
  Tenthanhcai: string | null | undefined;
  Madvql: string | null | undefined;
  Tencongty: string | null | undefined;
  Truyentaidien: string | null | undefined;
  Capda: string | null | undefined;
  Sohieu: string | null | undefined;
  Sohuu: string | null | undefined;
  Ngaylapdat: Date | null | undefined;
  Ngayvh: Date | null | undefined;
  Thuoctram: string | null | undefined;
  Tentram: string | null | undefined;
  Sohieubanve: string | null | undefined;
  Sododanhso: string | null | undefined;
  Cottenhienthi: string | null | undefined;
  Dahienthitrensd: string | null | undefined;
  Hienthiten: string | null | undefined;
  Hoatdong: string | null | undefined;
  Tthientai: string | null | undefined;
  Jsongeo: string | null | undefined;
  Maudong: string | null | undefined;
  Maucat: string | null | undefined;
  Ghichu: string | null | undefined;
}